[iOS] Move AIChatHistoryViewModelDelegate conformance to its own file
Extract MainViewController's AIChatHistoryViewModelDelegate conformance into
MainViewController+AIChatHistoryViewModelDelegate.swift, matching the existing
MainViewController+*.swift split. No behaviour change.
